+++ pigeon/pigeon.py
@@ -8,14 +8,15 @@ import argparse
 import internetarchive
 
 import tempfile
+import math
 import asyncio
+import requests
 
-from utils import (
-    get_paginated_data,
-    get_with_retry,
-    build_doi,
-    get_datacite_metadata
-)
+from typing import Tuple, Dict
+from ratelimit import sleep_and_retry
+from ratelimit.exception import RateLimitException
+from datacite import DataCiteMDSClient
+
 import settings
 import zipfile
 import bagit
@@ -70,7 +71,9 @@ def create_zip_data(temp_dir):
     return zip_data
 
 
-def upload(guid, zip_data, ia_access_key, ia_secret_key, collection_name=None):
+def upload(guid, zip_data, ia_access_key, ia_secret_key, collection_name=settings.OSF_COLLECTION_NAME):
+    assert isinstance(ia_access_key, str), 'Internet Archive access key was not passed to pigeon'
+    assert isinstance(ia_secret_key, str), 'DaInternet Archive secret key not passed to pigeon'
     session = internetarchive.get_session(
         config={
             's3': {
@@ -172,6 +175,105 @@ def pigeon(
 
         metadata = get_metadata(temp_dir, 'registraton.json')
         modify_metadata_with_retry(ia_item, metadata)
+
+
+def build_doi(guid):
+    return settings.DOI_FORMAT.format(prefix=settings.DATACITE_PREFIX, guid=guid)
+
+
+def get_datacite_metadata(doi, datacite_username, datacite_password, datacite_prefix):
+    assert isinstance(datacite_password, str), 'Datacite password not passed to pigeon'
+    assert isinstance(datacite_username, str), 'Datacite username not passed to pigeon'
+    assert isinstance(datacite_prefix, str), 'Datacite prefix not passed to pigeon'
+    client = DataCiteMDSClient(
+        url=settings.DATACITE_URL,
+        username=datacite_username,
+        password=datacite_password,
+        prefix=datacite_prefix,
+    )
+    return client.metadata_get(doi)
+
+
+@sleep_and_retry
+def get_with_retry(
+        url,
+        retry_on: Tuple[int] = (),
+        sleep_period: int = None,
+        headers: Dict = None) -> requests.Response:
+
+    if not headers:
+        headers = {}
+
+    if not settings.OSF_THROTTLE_ENABLED:
+        assert settings.OSF_BEARER_TOKEN, 'must have OSF_BEARER_TOKEN set to disable throttle'
+        headers['Authorization'] = settings.OSF_BEARER_TOKEN
+
+    resp = requests.get(url, headers=headers)
+    if resp.status_code in retry_on:
+        raise RateLimitException(
+            message='Too many requests, sleeping.',
+            period_remaining=sleep_period or int(resp.headers.get('Retry-After') or 0)
+        )  # This will be caught by @sleep_and_retry and retried
+
+    return resp
+
+
+@sleep_and_retry
+def put_with_retry(
+        url: str,
+        data: bytes,
+        headers: dict = None,
+        retry_on: Tuple[int] = (),
+        sleep_period: int = None) -> requests.Response:
+
+    if headers is None:
+        headers = {}
+
+    if not settings.OSF_THROTTLE_ENABLED:
+        assert settings.OSF_BEARER_TOKEN, 'must have OSF_BEARER_TOKEN set to disable throttle'
+        headers['Authorization'] = settings.OSF_BEARER_TOKEN
+
+    resp = requests.put(url, headers=headers, data=data)
+    if resp.status_code in retry_on:
+        raise RateLimitException(
+            message='Too many requests, sleeping.',
+            period_remaining=sleep_period or int(resp.headers.get('Retry-After') or 0)
+        )  # This will be caught by @sleep_and_retry and retried
+
+    return resp
+
+
+async def get_pages(url, page, result={}):
+    url = f'{url}?page={page}'
+    resp = get_with_retry(url, retry_on=(429,))
+    result[page] = resp.json()['data']
+    return result
+
+
+async def get_paginated_data(url):
+    data = get_with_retry(url, retry_on=(429,)).json()
+
+    tasks = []
+    is_paginated = data.get('links', {}).get('next')
+
+    if is_paginated:
+        result = {1: data['data']}
+        total = data['links'].get('meta', {}).get('total') or data['meta'].get('total')
+        per_page = data['links'].get('meta', {}).get('per_page') or data['meta'].get('per_page')
+
+        pages = math.ceil(int(total) / int(per_page))
+        for i in range(1, pages):
+            task = get_pages(url, i + 1, result)
+            tasks.append(task)
+
+        await asyncio.gather(*tasks)
+        pages_as_list = []
+        # through the magic of async all our pages have loaded.
+        for page in list(result.values()):
+            pages_as_list += page
+        return pages_as_list
+    else:
+        return data
 
 
 if __name__ == '__main__':
