+++ pigeon/settings/defaults.py
@@ -2,17 +2,17 @@
 
 DOI_FORMAT = '{prefix}/fk2osf.io/{guid}'
 
-DATACITE_USERNAME = ''
-DATACITE_PASSWORD = ''
+DATACITE_USERNAME = None
+DATACITE_PASSWORD = None
 DATACITE_PREFIX = '10.17605'  # Datacite's prod DOI prefix
 DATACITE_URL = 'https://mds.datacite.org/'
 
 PAGE_SIZE = 100
 OSF_COLLECTION_NAME = 'cos-dev-sandbox'
-OSF_BEARER_TOKEN = ''
+OSF_BEARER_TOKEN = None
 
-IA_ACCESS_KEY = ''
-IA_SECRET_KEY = ''
+IA_ACCESS_KEY = None
+IA_SECRET_KEY = None
 
 OSF_API_URL = 'https://api.staging.osf.io/'
 OSF_FILES_URL = 'https://files.us.staging.osf.io/'
