--- settings/__init__.py
@@ -1,11 +1,0 @@
-import warnings
-
-from .defaults import *  # noqa
-
-try:
-    from .local import *  # noqa
-except ImportError:
-    warnings.warn(
-        'No settings file found. Did you remember to '
-        'copy local-dist.py to local.py?', ImportWarning,
-    )
